requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aioprocessing==2.0.1
aiosignal==1.3.1
altair==5.3.0
analytics-python==1.2.9
annotated-types==0.7.0
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.1.1
backoff==2.2.1
bc-detect-secrets==1.5.9
black==24.4.2
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.6
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
deepeval==0.21.45
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
docx2txt==0.8
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
evaluate==0.4.2
exceptiongroup==1.2.1
execnet==2.1.1
executing==2.0.1
Faker==24.14.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
flake8==7.0.0
frozenlist==1.4.1
fsspec==2024.3.1
fuzzysearch==0.7.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
gql==3.5.0
graphql-core==3.2.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
importlib-metadata==7.0.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
isort==5.13.2
janus==1.0.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json_repair==0.17.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.60
language_data==1.2.0
litellm==1.37.20
llm-guard==0.3.13
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.30.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.37
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.8.2
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.2
pysbd==0.3.4
pytest==8.2.1
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytrec_eval==0.5
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ragas==0.1.8
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.4.28
replicate==0.26.0
requests==2.32.2
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.5
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
sentry-sdk==2.2.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
streamlit==1.35.0
structlog==24.1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tonic-validate==6.0.0
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.39.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
unidiff==0.7.5
uritemplate==4.1.1
urllib3==2.2.1
uv==0.1.45
uvicorn==0.30.0
uvloop==0.19.0
wandb==0.17.0
wasabi==1.1.2
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.3.4
weave==0.50.3
-e git+https://github.com/ash0ts/snowflake-arctic-weave-demo.git@ff51f81382a7e12c2f8b5816b49f935f736aea26#egg=weave_example_demo
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2
